feat: image cache기능 추가

+++ OpenMarket/OpenMarket/Model/JSONModel/ProductDetail.swift
@@ -11,8 +11,8 @@ struct ProductDetail: Decodable {
     let name: String
     let thumbnail: String
     let currency: Currency
-    let price: Int
-    let bargainPrice: Int
+    let price: Double
+    let bargainPrice: Double
     let discountedPrice: Int
     let stock: Int
     let createdAt: String
